chore: rename integration tests e2e
