+++ packages/client/e2e/step-definitions/evaluation.spec.ts
@@ -1,30 +1,46 @@
 import { defineFeature, loadFeature } from 'jest-cucumber';
-import { JsonValue, JsonObject, EvaluationDetails, EvaluationContext, ResolutionDetails, StandardResolutionReasons } from '@openfeature/shared';
-import { OpenFeature } from './../../src';
+import {
+  JsonValue,
+  JsonObject,
+  EvaluationDetails,
+  EvaluationContext,
+  ResolutionDetails,
+  StandardResolutionReasons,
+} from '@openfeature/shared';
+import { OpenFeature, ProviderEvents } from '../../';
 // load the feature file.
-const feature = loadFeature('packages/server/integration/features/evaluation.feature');
+const feature = loadFeature('packages/client/e2e/features/evaluation.feature');
 
 // get a client (flagd provider registered in setup)
 const client = OpenFeature.getClient();
 
-const givenAnOpenfeatureClientIsRegisteredWithCacheDisabled = (given: (stepMatcher: string, stepDefinitionCallback: () => void) => void) => {
+const givenAnOpenfeatureClientIsRegisteredWithCacheDisabled = (
+  given: (stepMatcher: string, stepDefinitionCallback: () => void) => void
+) => {
   // TODO: when the FlagdProvider is updated to support caching, we may need to disable it here for this test to work as expected.
   given('a provider is registered with cache disabled', () => undefined);
 };
 
 defineFeature(feature, (test) => {
+  beforeAll((done) => {
+    client.addHandler(ProviderEvents.Ready, () => {
+      done();
+    });
+  });
+
+  afterAll(() => {
+    OpenFeature.close();
+  });
 
   test('Resolves boolean value', ({ given, when, then }) => {
     let value: boolean;
-    let flagKey: string;
 
     givenAnOpenfeatureClientIsRegisteredWithCacheDisabled(given);
 
     when(
       /^a boolean flag with key "(.*)" is evaluated with default value "(.*)"$/,
-      async (key: string, defaultValue: string) => {
-        flagKey = key;
-        value = await client.getBooleanValue(flagKey, defaultValue === 'true');
+      (key: string, defaultValue: string) => {
+        value = client.getBooleanValue(key, defaultValue === 'true');
       }
     );
 
@@ -35,15 +51,13 @@ defineFeature(feature, (test) => {
 
   test('Resolves string value', ({ given, when, then }) => {
     let value: string;
-    let flagKey: string;
 
     givenAnOpenfeatureClientIsRegisteredWithCacheDisabled(given);
 
     when(
       /^a string flag with key "(.*)" is evaluated with default value "(.*)"$/,
-      async (key: string, defaultValue: string) => {
-        flagKey = key;
-        value = await client.getStringValue(flagKey, defaultValue);
+      (key: string, defaultValue: string) => {
+        value = client.getStringValue(key, defaultValue);
       }
     );
 
@@ -54,34 +68,31 @@ defineFeature(feature, (test) => {
 
   test('Resolves integer value', ({ given, when, then }) => {
     let value: number;
-    let flagKey: string;
 
     givenAnOpenfeatureClientIsRegisteredWithCacheDisabled(given);
 
     when(
       /^an integer flag with key "(.*)" is evaluated with default value (\d+)$/,
-      async (key: string, defaultValue: string) => {
-        flagKey = key;
-        value = await client.getNumberValue(flagKey, Number.parseInt(defaultValue));
+      (key: string, defaultValue: number) => {
+        value = client.getNumberValue(key, defaultValue);
       }
     );
 
-    then(/^the resolved integer value should be (\d+)$/, (expectedValue: string) => {
-      expect(value).toEqual(Number.parseInt(expectedValue));
+    then(/^the resolved integer value should be (\d+)$/, (expectedStringValue: string) => {
+      const expectedNumberValue = parseInt(expectedStringValue);
+      expect(value).toEqual(expectedNumberValue);
     });
   });
 
   test('Resolves float value', ({ given, when, then }) => {
     let value: number;
-    let flagKey: string;
 
     givenAnOpenfeatureClientIsRegisteredWithCacheDisabled(given);
 
     when(
       /^a float flag with key "(.*)" is evaluated with default value (\d+\.?\d*)$/,
-      async (key: string, defaultValue: string) => {
-        flagKey = key;
-        value = await client.getNumberValue(flagKey, Number.parseFloat(defaultValue));
+      (key: string, defaultValue: string) => {
+        value = client.getNumberValue(key, Number.parseFloat(defaultValue));
       }
     );
 
@@ -92,37 +103,31 @@ defineFeature(feature, (test) => {
 
   test('Resolves object value', ({ given, when, then }) => {
     let value: JsonValue;
-    let flagKey: string;
-
     givenAnOpenfeatureClientIsRegisteredWithCacheDisabled(given);
 
-    when(/^an object flag with key "(.*)" is evaluated with a null default value$/, async (key: string) => {
-      flagKey = key;
-      value = await client.getObjectValue(flagKey, {});
+    when(/^an object flag with key "(.*)" is evaluated with a null default value$/, (key: string) => {
+      value = client.getObjectValue<JsonValue>(key, {});
     });
 
     then(
       /^the resolved object value should be contain fields "(.*)", "(.*)", and "(.*)", with values "(.*)", "(.*)" and (\d+), respectively$/,
-      (field1: string, field2: string, field3: string, boolValue: string, stringValue: string, intValue: string) => {
+      (field1: string, field2: string, field3: string, boolVal: string, strVal: string, intVal: string) => {
         const jsonObject = value as JsonObject;
-        expect(jsonObject[field1]).toEqual(boolValue === 'true');
-        expect(jsonObject[field2]).toEqual(stringValue);
-        expect(jsonObject[field3]).toEqual(Number.parseInt(intValue));
+        expect(jsonObject[field1]).toEqual(boolVal === 'true');
+        expect(jsonObject[field2]).toEqual(strVal);
+        expect(jsonObject[field3]).toEqual(Number.parseInt(intVal));
       }
     );
   });
 
   test('Resolves boolean details', ({ given, when, then }) => {
     let details: EvaluationDetails<boolean>;
-    let flagKey: string;
-
     givenAnOpenfeatureClientIsRegisteredWithCacheDisabled(given);
 
     when(
       /^a boolean flag with key "(.*)" is evaluated with details and default value "(.*)"$/,
-      async (key: string, defaultValue: string) => {
-        flagKey = key;
-        details = await client.getBooleanDetails(flagKey, defaultValue === 'true');
+      (key: string, defaultValue: string) => {
+        details = client.getBooleanDetails(key, defaultValue === 'true');
       }
     );
 
@@ -138,15 +143,13 @@ defineFeature(feature, (test) => {
 
   test('Resolves string details', ({ given, when, then }) => {
     let details: EvaluationDetails<string>;
-    let flagKey: string;
 
     givenAnOpenfeatureClientIsRegisteredWithCacheDisabled(given);
 
     when(
       /^a string flag with key "(.*)" is evaluated with details and default value "(.*)"$/,
-      async (key: string, defaultValue: string) => {
-        flagKey = key;
-        details = await client.getStringDetails(flagKey, defaultValue);
+      (key: string, defaultValue: string) => {
+        details = client.getStringDetails(key, defaultValue);
       }
     );
 
@@ -162,15 +165,13 @@ defineFeature(feature, (test) => {
 
   test('Resolves integer details', ({ given, when, then }) => {
     let details: EvaluationDetails<number>;
-    let flagKey: string;
 
     givenAnOpenfeatureClientIsRegisteredWithCacheDisabled(given);
 
     when(
       /^an integer flag with key "(.*)" is evaluated with details and default value (\d+)$/,
-      async (key: string, defaultValue: string) => {
-        flagKey = key;
-        details = await client.getNumberDetails(flagKey, Number.parseInt(defaultValue));
+      (key: string, defaultValue: string) => {
+        details = client.getNumberDetails(key, Number.parseInt(defaultValue));
       }
     );
 
@@ -186,15 +187,13 @@ defineFeature(feature, (test) => {
 
   test('Resolves float details', ({ given, when, then }) => {
     let details: EvaluationDetails<number>;
-    let flagKey: string;
 
     givenAnOpenfeatureClientIsRegisteredWithCacheDisabled(given);
 
     when(
       /^a float flag with key "(.*)" is evaluated with details and default value (\d+\.?\d*)$/,
-      async (key: string, defaultValue: string) => {
-        flagKey = key;
-        details = await client.getNumberDetails(flagKey, Number.parseFloat(defaultValue));
+      (key: string, defaultValue: string) => {
+        details = client.getNumberDetails(key, Number.parseFloat(defaultValue));
       }
     );
 
@@ -210,13 +209,11 @@ defineFeature(feature, (test) => {
 
   test('Resolves object details', ({ given, when, then, and }) => {
     let details: EvaluationDetails<JsonValue>; // update this after merge
-    let flagKey: string;
 
     givenAnOpenfeatureClientIsRegisteredWithCacheDisabled(given);
 
-    when(/^an object flag with key "(.*)" is evaluated with details and a null default value$/, async (key: string) => {
-      flagKey = key;
-      details = await client.getObjectDetails(flagKey, {}); // update this after merge
+    when(/^an object flag with key "(.*)" is evaluated with details and a null default value$/, (key: string) => {
+      details = client.getObjectDetails(key, {}); // update this after merge
     });
 
     then(
@@ -248,7 +245,7 @@ defineFeature(feature, (test) => {
 
     when(
       /^context contains keys "(.*)", "(.*)", "(.*)", "(.*)" with values "(.*)", "(.*)", (\d+), "(.*)"$/,
-      (
+      async (
         stringField1: string,
         stringField2: string,
         intField: string,
@@ -262,12 +259,14 @@ defineFeature(feature, (test) => {
         context[stringField2] = stringValue2;
         context[intField] = Number.parseInt(intValue);
         context[boolField] = boolValue === 'true';
+
+        await OpenFeature.setContext(context);
       }
     );
 
-    and(/^a flag with key "(.*)" is evaluated with default value "(.*)"$/, async (key: string, defaultValue: string) => {
+    and(/^a flag with key "(.*)" is evaluated with default value "(.*)"$/, (key: string, defaultValue: string) => {
       flagKey = key;
-      value = await client.getStringValue(flagKey, defaultValue, context);
+      value = client.getStringValue(flagKey, defaultValue);
     });
 
     then(/^the resolved string response should be "(.*)"$/, (expectedValue: string) => {
@@ -275,7 +274,8 @@ defineFeature(feature, (test) => {
     });
 
     and(/^the resolved flag value is "(.*)" when the context is empty$/, async (expectedValue) => {
-      const emptyContextValue = await client.getStringValue(flagKey, 'nope', {});
+      await OpenFeature.setContext({});
+      const emptyContextValue = client.getStringValue(flagKey, 'nope', {});
       expect(emptyContextValue).toEqual(expectedValue);
     });
   });
@@ -289,10 +289,10 @@ defineFeature(feature, (test) => {
 
     when(
       /^a non-existent string flag with key "(.*)" is evaluated with details and a default value "(.*)"$/,
-      async (key: string, defaultValue: string) => {
+      (key: string, defaultValue: string) => {
         flagKey = key;
         fallbackValue = defaultValue;
-        details = await client.getStringDetails(flagKey, defaultValue);
+        details = client.getStringDetails(flagKey, defaultValue);
       }
     );
 
@@ -318,10 +318,10 @@ defineFeature(feature, (test) => {
 
     when(
       /^a string flag with key "(.*)" is evaluated as an integer, with details and a default value (\d+)$/,
-      async (key: string, defaultValue: string) => {
+      (key: string, defaultValue: string) => {
         flagKey = key;
         fallbackValue = Number.parseInt(defaultValue);
-        details = await client.getNumberDetails(flagKey, Number.parseInt(defaultValue));
+        details = client.getNumberDetails(flagKey, Number.parseInt(defaultValue));
       }
     );
